fix(franchise): move franchise_error to core to resolve L2-imports-L1 contract violation
franchise/_api.py was importing from autoskillit.pipeline.gate, violating the
import-linter contract 'L2 franchise does not import L3 or most L1 peers'.

- Move franchise_error() body from pipeline/gate.py to core/_type_helpers.py,
  where it belongs (only needs L0 FRANCHISE_ERROR_CODES and stdlib json)
- pipeline/gate.py imports and re-exports franchise_error from core (noqa F401)
  so all existing callers via autoskillit.pipeline remain unaffected
- core/__init__.py re-exports franchise_error alongside FranchiseErrorCode
- franchise/_api.py imports franchise_error from autoskillit.core only
- Replace 3 gate_error_result() calls in franchise/_api.py with franchise_error()
  using FRANCHISE_MANIFEST_MISSING (infrastructure not initialized)
