+++ src/autoskillit/core/__init__.py
@@ -164,6 +164,7 @@ from .types import (
     WriteExpectedResolver,
     extract_path_arg,
     extract_skill_name,
+    franchise_error,
     resolve_target_skill,
     session_type,
     truncate_text,
@@ -220,6 +221,7 @@ __all__ = [
     "WriteExpectedResolver",
     "extract_path_arg",
     "extract_skill_name",
+    "franchise_error",
     "resolve_target_skill",
     "PackDef",
     "PACK_REGISTRY",
